enable LUA option for nginx
currently, mod_rewrite hack is used to switch maintenance mode on/off.
similar hack is needed for turtle, where nginx is used as reverse
proxy. however, default nginx cannot do the same thing mainly because
"if" does not have "<" or ">" to compare numbers. LUA is enabled in the
hope that we can do it in lua.